easyCollector/requirements.txt
# ============================================================================
# Easy Collector - Data Collection Service Requirements
# ============================================================================
# IMPORTANT: This is the requirements.txt for Easy Collector service ONLY
# DO NOT confuse this with the parent directory's requirements.txt which is
# for "Easy ORB Strategy" - they are DIFFERENT files with DIFFERENT purposes!
#
# Location: easyCollector/requirements.txt
# Parent requirements.txt: ../requirements.txt (for Easy ORB Strategy)
#
# FastAPI + GCP Services for Market Data Collection
# Optimized for deployment with Easy ORB Strategy and Easy 0DTE Strategy
# Minimal dependencies to reduce deployment size
# ============================================================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Cloud Platform (minimal set for Firestore only)
google-cloud-firestore==2.14.0

# Data Processing (aligned with Easy ORB Strategy versions)
pandas==2.1.4
numpy==1.26.3

# HTTP Client (for market data APIs)
httpx==0.26.0
aiohttp==3.9.1

# Market Data (E*TRADE fallback)
yfinance==0.2.32

# Date/Time Handling
pytz==2024.1
python-dateutil==2.8.2

# Holiday Calculations (shared with Easy ORB Strategy)
# Note: Can use shared modules/dynamic_holiday_calculator.py when deployed together
holidays==0.36

# Logging & Monitoring
python-json-logger==2.0.7

# Testing (development only)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
# google-cloud-firestore-emulator==0.0.2  # Not available for Python 3.11, use local emulator instead

# Environment & Config
python-dotenv==1.0.0

# Utilities
tenacity==8.2.3  # Retry logic
